Add missing PinDescription fields to fix memory layout mismatch.
Code that used the Core headers had the wrong layout, so we were reading random
areas of memory. This makes the layout of g_APinDescription the same no matter
which header a source file includes.

This is ugly and long term I think we need one foundation to build upon.